requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1

# Database
motor>=3.1.1
pymongo>=4.3.3

# Vector Database
pinecone-client>=2.2.1
chromadb>=0.4.6

# Embedding and LLM
openai>=0.27.4
langchain>=0.0.200
langchain-openai>=0.0.1
tiktoken>=0.4.0

# Utilities
numpy>=1.24.2
python-dotenv>=1.0.0
httpx>=0.24.0
tenacity>=8.2.2

# Connectors
google-auth>=2.16.2
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.86.0
slack-sdk>=3.21.3
notion-client>=2.0.0
